Use custom AuthInterface provider for Member JWT
This avoids the need to implement Authenticatable for the
Member model, which would be useless, as they would all just
be dummy methods to satisfy the interface requirements. This
application does not do any authentication itself (that is handled
by outside sources by e.g. Google and Shibboleth), only verification
of said authentication.