Support passing a callback to /auth/{provider}
Add support for sending a callback to /auth/{provider}.
Example flow:

GET {host}/api/v1/auth/google&callback={callback_url}
Callback is encoded in a state object and is sent along with a redirect to the
OAuth2 provider. Requesting entity authenticates and is redirected back to
/auth/google/callback with standard OAuth2 parameters. If authentication is
verified, the state will be decoded and an attempt will be made to extract a
callback. If one is found, a redirect response will be executed using the
callback url with the generated authentication token sent in a 'token'
parameter. Otherwise, a JSON object will be sent back with a single key,
'token'.

Closes #29