Add service level authentication: GET /auth/token
Three parameters are required: provider, id, secret.
Provider is meant to be an external provider such as slack or github. The id
field refers to the external id of these services that the target member is
associated with. Finally, the secret field authenticates the trusted service
with the API.

If success in finding, the endpoint will return a JSON object with a single
key: 'token' which has a JWT as a value. Otherwise, either a 404 or a 401
response will be generated.

Closes #31.